Allowlist TaintInTaintThis behaviour
Summary: Previous behaviour was that we add TaintInTaintThis behaviour to any method that doesn't have code and didn't match the denylist. This was causing false positives as in reality only object setters will taint Argument(0). For example cases like String::replace, Map::put should have TaintInTaintThis behaviour while String::equals should not. This diff is not perfect but should address 95%+ of all setters and in doing so eradicate a fair portion of false positives where setters are not used.

Reviewed By: yuhshin-oss

Differential Revision: D34478587

fbshipit-source-id: 2c53bb931783338d16cf417b1683db1c905be1ca