Introduce strict JoinOverrideThreshold for Android methods
Summary: Following overrides of Android or other third party library methods commonly leads to false positives as a typed base instance would resolve to all instances of the method where this is most often not the actual behaviour in code. Therefore applying a stricter limit to Android specific code to limit these patterns.

Reviewed By: SimranVirk

Differential Revision: D34527444

fbshipit-source-id: b589ecc3d9b12175f94263184650f96951779cae