string_views, string_view everwhere!
Summary:
This adds some extra copies, now highly visible via new `str_copy` helper function.
But it turns out that it also immediately removes a bunch of (formerly implicit) copies.
Furthermore, I already cleaned up / converted to std::string_view various code in the vicinity of string uses to minimize any overhead.

We can work on reducing the remaining `str_copy` incrementally.

The main purpose of this diff is to prepare for the next ones that save memory.

This is a behavior-preserving change.

Reviewed By: yuxuanchen1997

Differential Revision: D35060908

fbshipit-source-id: a16aa2a17a75fd4949f0722eb215d36cd81d76a9