TaintV1: Expose `Frame` instead of `FrameSet` (via TaintV1::frames_iterator)
Summary:
Part of the changes for moving Taint structure from `Kind -> Callee -> Position -> Frames` to `Callee -> Position -> Kind -> Frames`

`FrameSet` is an implementation detail of `TaintV1` and should not be exposed as part of the `Taint` API.

It is currently exposed in `TaintV1`'s iterator. Changing the iterator to operate on `Frame` instead. Done using `FlattenIterator`.

___

Reviewed By: SimranVirk

Differential Revision: D34573812

fbshipit-source-id: b83a3e487ab826a4149f97bf6ce846cfec3034b3