TaintV2: CalleePortFrames: propagate
Summary:
**Recall:** TaintV2 structure in the process of changing from `Callee -> CallPosition -> Kind -> Frame` to `Callee -> CallPosition -> **CalleePort** -> Kind -> Frame`.

`CalleePortFrames::propagate` implemented here looks very similar to `CallPositionFrames::propagate`. The main difference is that "callee_port" can be assumed to be the same in all frames.

This simplifies the implementation slightly in that we do not need to partition frames by callee port in here. Everything else is copied from `CallPositionFrames`. These will be removed from that class later when we update it to use `CalleePortFrames`.

Reviewed By: SimranVirk

Differential Revision: D35729750

fbshipit-source-id: 9263fe1c0bd7cc63a7f2ca8bdeed15bec2d61f74