Improve sample app document + bug fix
Summary:
The parameters to pass in to Mariana Trench rely in the documentation on default values. This is causing confusion and missing configuration. E.g. When testing this out not specifying the model generator search paths will miss models not included in the build and forgo lifecycle methods.

Additionally we had 2 sample apps checked in so deleting the deprecated one and updating the new one to correct a bug

Reviewed By: SimranVirk

Differential Revision: D35776419

fbshipit-source-id: 679b2c05ac4b6a59d9a7a7d1600673541dbe6076