Json shim generator: Fix dropping arguments in ArtificialCallee.
Summary:
With shims, we can artificially invoke methods. For these artificial callees, we map the arguments from the original method to the artificial callees (either user defined or inferred).
However, we may *not* have all the arguments required to invoke these methods. Currently, in such cases, if the arguments are dropped before the end, the remaining ones are not mapped correctly as the arguments for `ArtificialCallee`s are stored as a list of registers. This diff fixes this by storing the arguments for `ArtificialCallee` as a map from position to register instead. Corresponding change is also made in `ShimTarget`. Also, the json specification for shims is changed to reflect this as follows:
```
"shim": {
    "callees": [
      {
.       "signature": "<exact signature of the method to call>",
        "parameters_map": {
          "Argument(0)": "Argument(<int>)", // Specifies the receiver
          "Argument(<int>)": "Argument(<int>)", ...
        }
      },
      ...
    ]
  }
```

Reviewed By: yuhshin-oss

Differential Revision: D35987030

fbshipit-source-id: a29c18e848b324df0342f05a81d0dd875d75dca9