TaintV2: JSON V2
Summary:
Add support for JSON V2 for TaintV2. This is represented as "0.2" in the metadata.json file. Downstream parsers will determine which format to use based on the version metadata.json.

Switching between V1 and V2 is controlled by a compile-time constant in Constants.h. As of this change, we are still on V1.

The constant will live until we are sure that V2 works well (couple of weeks) and downstream dependencies have been updated to support it.

Reviewed By: SimranVirk

Differential Revision: D36071179

fbshipit-source-id: 7b9eac3cf3690bfd25ffff9e2a76cfa680e4dffa