TaintV2: Update tests with json v2 format
Summary:
Switch to the new JSON format.

NOTE: For open source, if building Mariana Trench from source that includes this change, this is not compatible with the current latest version of sapp-cli (0.5.2). Will be bumping that version shortly. Meanwhile, fb-sapp needs to be built from source to work on the new format. Alternatively, switch the `k_is_legacy_output_version` to `true` to use the old output format.

Reviewed By: dkgi

Differential Revision: D36084330

fbshipit-source-id: ac388fb83145221fe221299620e8419d0917060c