Add via-file feature to improve SNR of category 4038
Summary: In category 4038 we want to detect User input flowing into a File. The problem is that this also detect User input -> File -> File. To filter out the latter flow we can add a feature to File flows and filter them out (meaning we reach the file sink through another file object).

Reviewed By: yuhshin-oss

Differential Revision: D36663100

fbshipit-source-id: df358bef849f245a3653be46a25c0279c09e98ae