Add feature for dependency graph traces with 5+ depth
Summary:
The dependency graph features are a heuristic to determine if a specific callsite is exported or not. These features are not always correct for example in cases like

```
function z() {
  y(1);
}

function y(a) {
  if (a == 2) {
    callable()
  }
}
```
here we'd consider callable() being in the dependency graph of z() while in reality z() never calls callable();

Therefore shorter depth have better SNR and using this features we can exclude 5+ depths more likely to contain the pattern shown above.

Reviewed By: SimranVirk

Differential Revision: D36665740

fbshipit-source-id: 3ad33d56ffd008deebb33c14868e267a5237517a