Add information for the new handles to `Issue`s
Summary:
This added `sink_index`, `callee` and `callee_port` to a struct which should be held by each `Issue`. There were a couple other ways to do things:

1. Add these three directly to the issue rather than holding a struct.
**Downside:** iI we require more info to be stored for handles later on, we don't need to change issues' callsites etc. Also, we need to check equality of all the 'handle information' in `Issue`'s join/leq/etc methods. Holding them in a struct means we only need one assert rather than 3 (alternatively, encapsulate all these asserts in a helper method, then there's not much difference whether we hold the fields directly or within the struct).

2. Store the string handle in an issue rather than all the individual parts of the handle
**Downside:** It seems less safe for equality checks during join/leq etc.

3.  Store the position (and/or rule) within the `IssueHandleInfo` as well.
I can't see any downside/upside here. Open to ideas.

I kept the change to Issue's json format in the next diff, since this changes the results from all our tests. For that as well, there are a few options which depend on this diff.

Reviewed By: yuhshin-oss

Differential Revision: D35874413

fbshipit-source-id: 7e0d5a62301e05e35299759c1737e9c4c72eb078