Do not compute array allocation kind if unused
Summary:
`UnusedKinds`, as the name describes, computes kinds that are not used in any rules. It then removes those kinds from models that use them. This saves us from computing models that will never trigger an issue.

However, the "ArrayAllocation" kind is computed in Transfer.cpp rather than configured via models because it operates on the `new` operator/instruction rather than a method/field. This means it is always being computed.

This change adds a check in Transfer.cpp for whether the kind is used before we compute it.

Reviewed By: anwesht

Differential Revision: D37408354

fbshipit-source-id: 00a4a96c960aad1d5a0ff22fc3dacab1c90a911b