Skip shim generation for calls within libraries
Summary:
Specifying a shim generator with: shimmed-method: `foo()` and shim-target: `bar()` introduces an artificial call to `bar()` on all callsites of `foo()`. Currently we are adding these artificial calls for callers within libraries such as `Landroid/`, `Landroidx/`, `Lcom/google/` as well. This diff filters out these callsites.

Currently this is just a static list within the binary. We can consider making this configurable in the future if needed.

Reviewed By: yuhshin-oss

Differential Revision: D38218841

fbshipit-source-id: 899db4e8aa5ea1af8f618277332096bac3a6d439