Investigate memory increased memory usage
Summary:
We've noticed a gradual memory usage increase in mariana trench over the last month. The initial spike is after `mariana_trench:259` release on 07-04-2022 and gradually increases there after. This release includes:

| D37503014 (https://github.com/facebook/mariana-trench/commit/aa790cac77cefed7bfbb4ba46f403e9215de0971)	| [mariana trench][easy] Run reflection analysis even if the global type analysis is run|
| D37475028 (https://github.com/facebook/mariana-trench/commit/b3724cab0d724453b04fbfd36ecaabaefeb75a9e)	| [mariana trench][easy] Improve type unit test|
| D37545799	| [mariana trench][easy] Fix wiki entry for running on apps
| D37408354 (https://github.com/facebook/mariana-trench/commit/2d6e9ed0337762c0baf10b1621ca5273c1057d26)	| [MT] Do not compute array allocation kind if unused
| D37196279 (https://github.com/facebook/mariana-trench/commit/5479db683368454a1006bed884f32ec611bcb354)	| [MT] Add option to disable via-cast feature
| D37157237 (https://github.com/facebook/mariana-trench/commit/f882be470d5c42cc5a448d09d069c6734d43a484)	| [sparta] [ConstantAbstractDomain] use less memory
| D37405142 (https://github.com/facebook/mariana-trench/commit/de42773f68f4ceeea3668f2b836fc5481cc5b6b2)	| [MT] Added testcase for shims
| D37405143 (https://github.com/facebook/mariana-trench/commit/f996c671542c9106554d0d52a8e18ad02727e6bd)	| [MT] Resolve receiver mapping for lifecycle shims

The 2 big features landed during this time were *shims* and *global types*. It is expected to see an increase in memory usage with these features as addition of shims extends the callgraph and hence *more things are analyzed*. But both are configurable and we still see increased memory usage when disabling these features, which is weird.

Looked at the shims changes and saw that computing `MethodMappings` is pretty memory hungry. We do this twice (1 for shim + 1 for model generation).
This diff refactors `MethodMappings`:
- Move to separate file
- non-copyable
- uses the same instance and clears the data after both model and shim generation are complete.

Also:
- clears the generated shims.
- Adds RSS logging after each step.

 ---