Fix FNs: Join field models with access path reads.
Summary:
This diff fixes the field update via assignment bug described in the previous diff by joining the user defined taint with the existing taint information on the field instead of overwriting it.

Note: We also do a *strong* update in `analyze_sget()`  but that is fine as we do not handle writes to the static fields. We only implement `analyze_sput` so that we can define field sinks on them. Other propagations through static fields are not tracked. This is consistent with zoncolan.

Reviewed By: SimranVirk

Differential Revision: D38644447

fbshipit-source-id: 80600ed12d85cb0bf81c8e8d4b96dd25c976e7f9