Testcases for field sources/access paths inconsistencies
Summary:
# What are Field Sources?
Field sources allows us to define a *class field*  as a source of specified kind. This implies that all *accesses* of that field on any instance of that class *returns* a source. But we express taint flows at the granularity of methods (i.e. taint flows are specified at method boundaries on parameters (including this), or the return ports). So to specify fields as sources, we model the *field access* as a call to a special "field_callee" which *returns* the specified source.

# Existing Problem
Currently, ~~we do not propagate any *updates* to fields.~~ we do track writes to fields via `iput` instruction but when reading the field via `iget` instruction, we overwrite the taint (strong update) with the user defined field source instead of joining (weak update).

This diff adds testcases to  show how field sources and access paths interact + show the existing FN. Fix in next diff.

Reviewed By: SimranVirk

Differential Revision: D38644449

fbshipit-source-id: b7d713273880f40aafda3920f36d8c4d00bd282f