Move local_positions: Remove parsing of local_positions from Frame::from_json
Summary:
The plan is to move `Frame::local_positions_` into `CalleePortFrames::local_positions_`, which means parsing of that value should not be done in `Frame`.

Frames with the same (callee, callee_position, callee_port) should have the same set of local positions, so it should be parsed and stored at the `CalleePortFrames` level. Note that this has not been implemented yet.

There are currently no user-defined model generators that specify "local_positions".

Reviewed By: anwesht

Differential Revision: D38643305

fbshipit-source-id: cf98abffcd01b415b69305fd24dac63216780ad8