Construct Taint objects from TaintBuilder
Summary:
In every class where that is constructed from `Frame`, add the equivalent constructor for `TaintBuilder`. For now, both constructors live side-by-side and the code looks rather similar. The following class constructors are affected:

* Model/FieldModel
* Taint/CallerFrames/CallPositionFrames/CalleePortFrames

The goal is to update all call-sites to the `TaintBuilder` constructor and ultimately remove the constructors that rely on `Frame`.

Reviewed By: SimranVirk

Differential Revision: D38884119

fbshipit-source-id: 8b8d70a942e4dc11fa387bb96964e2d15d389d1f