Add `Taint::set_leaf_origins_if_empty`
Summary:
Add `Taint::set_leaf_origins_if_empty` API. It is intended to replace the call to `Frame::set_origins` in Model.cpp.

This is called in Model.cpp when a `Model` is instantiated with an actual `Method`. The "origins" field can only be set at this point because we may not know the `Method` that the `Model` applies to prior to this.

This API is to be used for moving away from operating on `Frame` in `Model`.

Reviewed By: SimranVirk

Differential Revision: D39124766

fbshipit-source-id: f50c7e9986f243864207cca9af1f88fec11ad228