Add TaintBuilder to support construction of Taint object
Summary:
Today, `Taint` objects are constructed by passing `Frame`s in. However, there are properties in `Frame` that we want to move out of it. For example, `local_positions` and `local_features` should be in `CalleePortFrames`. Other fields like callee, callee_port, etc. do not need to be in `Frame` either.

If these fields are to be moved out of `Frame`, we need another way construct the `Taint` object from our model generators and tests. The solution is to introduce a `TaintBuilder` that contains these properties which we can eventually remove from `Frame`. However, it isn't an abstract domain the way `Frame` and various objects contained in the `Taint` class are.

This diff introduces the class, subsequent ones will use it. Implementation heavily copied from `Frame`, with one additional field: `local_positions` (it was moved out to `CalleePortFrames` previously).

Reviewed By: SimranVirk

Differential Revision: D38884120

fbshipit-source-id: 7ed96f268c40acf9e714dcace5f3206003a36651