Add Call Effects Abstract Domain
Summary:
Effects are modeled using flows through an artificial port on "CallEffect" root.  An effect will always have the implied Root of "CallEffect" and each type of effect we add will be a "path" on this port. `CallEffectAbstractDomain` handles the mapping of an "effect" to the "access path" internally.

This diff adds:
- `CallEffect` class to represent a kind of effect.
- `CallEffectAbstractDomain` to represent the taint for each kind of effect.

Reviewed By: SimranVirk

Differential Revision: D39118328

fbshipit-source-id: a0ebb1aabd9c42ed7ca4c524e5fbdbe3cc707a85