Support call effects in the analysis
Summary:
Handle taint flows through the "CallEffect" port in the analysis.
This implements the callchain tracking effect. Basically, for every "invoke" of a method, the call-chain sink is propagated from the `callee` to the `caller` until it meets a call-chain source, at which point an issue is created.

This diff:
- Adds call effect sources and sinks to `Model`
- the call effect sources behave like parameter sources on the call effect port.
- `check_call_effect_flows()` creates issues when the call effect sink bubbles up to meet the call effect source.
- `apply_call_effects()` propagates the call effect sinks up in the call chain in the same way as inferred sinks.

Reviewed By: SimranVirk

Differential Revision: D39182914

fbshipit-source-id: e40e35108930fbf827cf30659db1d5bb3c1f11be