Explicitly model getArguments & requireArguments
Summary: Currently requireArguments is a source since we are able to analyze it calls getArguments (we have the source code for androidx). This does however make the propagation a bit implicit and the flow a hop longer. Being explicit about the propagation these methods should follow and marking them both as source is better in reasoning about the shimming of Fragments.

Reviewed By: SimranVirk

Differential Revision: D39177089

fbshipit-source-id: 97066800035c46b9bdf227c06b2ed8785dd9ef79