Don't add tito/taint-in-taint-this to abstract/interface methods that have no obscure overrides
Summary:
If an interface/abstract method has <40 overrides (i.e we can join all the overrides at callsites), and all overrides are not obscure, then the interface/abstract method should not be marked as obscure (since every callsite is calling one of the overrides)

However, if an obscure method interface/abstract method doesn't have any overrides at all we also keep the tito mode (otherwise we get FNs. Many android package methods like Uri;.toString etc are interface and obscure to the analysis with no overrides, but we want to retain propagations for that).

Reviewed By: anwesht

Differential Revision: D39401783

fbshipit-source-id: 62724c4ef5d3534799a654c02006a40e25825a8c