Store tree of input paths in artificial source CalleePortFrames
Summary:
The input paths will be used to infer more precise propagations and sinks. They should be helpd only in artificial sources.

This change just introduces the field. Functionality and tests are added in the next diff.
The input paths are added to `CalleePortFrames` rather than `Frame`, since the input paths for a given artificial source root will be the same (this is because all artificial source frames with the same callee port root get joined, and so if the input paths were held within frames, they would get joined too)

Reviewed By: arthaud

Differential Revision: D40133400

fbshipit-source-id: f4fde44ac7feeb31a5df1674dd789b9543af7628