Remove callee port appending logic
Summary:
This removes all special-case handling of callee ports for artificial sources. The callee ports for artificial sources can now only hold a root (path must be empty). Paths are held in the `input_paths` field added in `CalleePortFrames`

Previously, the `input_paths` for `CaleePortFrames` were automatically set using the path from the `callee_port`, and we could continue doing that, but this would require `CalleePortFrames` to internally ensure that the path of the callee port is empty. This seems more error prone than just passing in the arguments to construction with this already ensured and then we can just assert that `input_paths` are bottom for non artificial sources and that the `callee_port` has no path for artificial sources

Reviewed By: arthaud

Differential Revision: D40375976

fbshipit-source-id: a67142e58b2c944665bc72022aa856c323475d3d