Remove FragmentUserInput
Summary: Marking Fragments as sources is an outdated decision made because there was no way to propagate taint to Fragments. Currently using shims on setArguments we have better and more accurate tools to propagate taint to Fragments and while not perfect the current SNR on FragmentUserInput is sufficiently problematic that its not worth keeping around and better focused on improving false negatives like T135580498.

Reviewed By: dkgi

Differential Revision: D41188177

fbshipit-source-id: db01fc90af095008eb0d47f74ea5d094dc8f4def