Rewrite ClassProperties to be component specific
Summary:
Because ClassProperties places all components in a single set/map a Receiver can be regarded as exported through an Activity and similar problems arise with Services. To improve the situation Im making sure that a source kind can only be considered exported by its equivalent component
- ActivityUserInput -> Activity
- ProviderUserInput -> Provider
- ReceiverUserInput -> Receiver
- ServiceUserInput && ServiceAIDLUserInput -> Service

The downside is that we hardcode the JSON UserInput kind in the binary.

In the process I have also removed via-child-exposed and merged this within via-caller-exported. This could be improved later on but the bulk should be solid.

Reviewed By: arthaud

Differential Revision: D41376354

fbshipit-source-id: 69a0857946768bb3dd61d5b8c93cd31cacfb6a79