Add new any_index type: value_of root
Summary:
This diff introduces a new named constructor for `PathElement` which takes a `Root` and expects it to be of kind `Argument`. This will create a `PathElement` of kind `AnyIndex` which, instead of a nullptr, stores the argument position as a string. At the callsite, this argument position will be resolved to the constant string which will then be used to construct the resolved `Path` as follows:
- If the argument position consists of a constant string: create a PathElement of kind `Index` with the value set to the resolved string
- if the argument position cannot be resolved to a constant string: create a PathElement of kind `AnyIndex` with nullptr.

Reviewed By: arthaud

Differential Revision: D41393385

fbshipit-source-id: 2725cf52390deb79061370db7e38ac9879d247da