Add ability to freeze models
Summary:
This diff adds the ability to "freeze" a section of the method's model (generations, parameter sources, sinks, and propagations). This basically means that these models are "higher priority" and supersedes other models that are not "frozen". Implementation is analogous to `Model::Modes`
- enum `FreezeKind` == `Mode` ==> represents which part of the model to "freeze".
- `Frozen` == `Modes` ==> Flags of `FreezeKind`

# Why do we need this?
In the following diff (), we add precise models for dictionary-like api for Intent and Bundle types and we will be adding more of these. The problem when testing these changes was that there are multiple model generators that match these methods ("low level android apis"?). When multiple models are discovered for the same method, we join them and end up with less precise/over-approximated models. This diff *overcomes* this by adding the ability to mark the more precise models as higher priority than other "capture all" style model generators eg. taint-in-taint-out models. We will still need the "capture all" models as we won't be maintaining models for every android api.

Initially, I just added a "freeze" as a mode but we do need a more finer grained ability to "freeze". Eg:
- There can be a precise model for propagation of a method but we might still want to add generations or sinks to the model from a different generator.
- There are "capture all" style model generators that add features which are then used in filters and hence need to be joined in.

# Semantics of freezing
- This diff allows us to "freeze" models of `FreezeKind`: "generations", "parameter_sources", "sinks", and "propagation".
- LessOrEqual comparison:
  - A "frozen" domain is greater than any domain that is not, regardless of the contents.
  - Comparing two "frozen" domains results in the regular leq comparison.
- Join:
  - No change when joining two domains both of which are *not* frozen.
  - Joining a "frozen" domain with a domain that is *not* results in the "frozen" domain.
  - Joining two different "frozen" domains results in the regular join.

Reviewed By: arthaud

Differential Revision: D41875728

fbshipit-source-id: cf9bf9447c9a5b0544d6eab83bb6ba40fae53047