create RAII device to manage RedexContext*
Summary: We have a `g_redex` pointer that clients are supposed to set/manage the lifetime of. We previously did so via `new`/`delete` in tests and in the main Mariana Trench. This isn't great, so let's introduce an RAII device called GlobalRedexContext that manages `g_redex` for us, and use it where we used to have `new ` and `delete `.

Reviewed By: arthaud

Differential Revision: D42461537

fbshipit-source-id: 3b360f01ba0bfacc99e456b022b1afd843a79c39