add a LOG_IF_INTERACTIVE macro and use it to avoid printing intemediate states in Sandcastle
Summary:
Context: Noticed that our logs were noisy for remote runs due to us unconditionally printing intermediate state that's mostly useful for a user who's waiting on the run.

The LOG_IF_INTERACTIVE macro checks if the stderr stream is a tty, and only logs if so - this should allow users to still see the *visceral* progress while making remote logs easier to triage.

Reviewed By: arthaud

Differential Revision: D42861733

fbshipit-source-id: d5cf1f2422045099c2ba3871b0336a96a698bdbd