Distinguish propagated and declared taint
Summary:
Context: Other taint analysis tools such as Pysa and Zoncolan associate a variant with frames that distinguish between declared taint, the origins of taint, and propagated taint. This allows these tools to avoid emitting frames for declared sources.

In Mariana Trench, we previously didn't have such a distinction, which means that the tool in general would join declared and user- taint. This is problematic when inferred and declared taint are joined - if we wanted to treat the origin taint as having no callees, there would be no way to distinguish origins from declared taint.

This diff adds a `CallInfo` enum that distinguishes inferred and declared taint. There are 3 kinds of taint:

* Declaration: This indicates taint that originate from a user-declared model, and should not be present in final traces.
* Origin: These frames are created by propagating declaration taint, and serve as the leaves of taint. The callees are scrubbed to ensure the traces don't attempt to go to the declaration.
* CallSite: Any frame that's propagated from an origin or another CallSite will be a call site.

(Note that CRTEX frames are special, and that they're treated as origins to ensure we keep them in the traces in models).

The key problem here for correctness is avoiding the merging of declared and origin/callsite taint, so the taint structure is also changed. Instead of a callee -> frame map, we now have a (callee * CallInfo) -> frame map, meaning that by design only the taint with the same info will get joined. This brings the taint structure closer to Zonk/Pysa.

Design decisions:
* Keeping Origin: We could also consider a representation of only Declaration/Inferred, with a distance of 0 inferring that the current frame is an origin of taint as well, and reducing the size of the taint map. This diff chooses to keep Origin as it's consistent with the rest of the codebases and easier to reason about compared to an implicit property, but it does come with overhead. We might want to reconsider this in the future if perf is a concern.
* Reducing all distances by 1: Note that this diff changes the `distance` field's values for almost all traces (see tests). Now that we're emitting 1 fewer leaf, the distance computation is brought in line with other tools for the same issues. This might have performance ramifications, and we may need to change our runs to reduce the limit by 1 after this diff lands.
* Not emitting the CallInfo in traces: I believe that we can construct the same traces without the extra information since we already carry distance info, so I chose to hide it from the JSON. Others may disagree, happy to revisit this one.

Reviewed By: arthaud

Differential Revision: D42720937

fbshipit-source-id: b0120addafb72776f2adcfb049ef5cc0dfd80f2d