Remove calls from Origin frames in the non-crtex case
Summary:
Context: We now distinguish between declared, origin and call sites for frames (see D42720937). The intent behind that diff was  to overall display one fewer leaf, but since the diff was getting unwieldy, I chose to split this diff off.

This diff: Stop emitting "call"'s at origin frames with the exception of CRTEX frames, where we still emit the call. This will mean that we now view origins as leaves. See the results in the diff :)

Reviewed By: arthaud

Differential Revision: D42815219

fbshipit-source-id: dad331ce5306c080eed8267838395f630d82a42c