Add a special call info for backward taint [backward analysis 1/n]
Summary:
Once we have a backward analysis, we will need to introduce taint for the local return and `this`.
We could use `Declaration` as the call info, although that is better used for user-declared taint.
Let's introduce a new call info `Propagation` that is only used for the backward analysis.

Reviewed By: anwesht

Differential Revision: D43014403

fbshipit-source-id: efc35ddf6757f03a85d8e91df5698876e478bff4