Better representation for local return and receiver kinds [backward analysis 5/n]
Summary:
This diff implements a new representation for a local return and receiver kinds.
These kinds will be used in a backward analysis to infer propagations.

The new design extends the `Kind` base class with a `LocalKind` class which represents any local "root". On top of that, we have `LocalReturnKind` and `LocalArgumentKind`.

The benefit of this approach is that we can treat both local kinds using `LocalKind` and `LocalKind::root()`. For instance, in the future, we can treat the inference of propagations in the same way. We can also easily extend it to infer propagations from an argument to another (non-this) argument. I could also see a future where we use `Taint` to represent propagations, and this would allow it.

The downside is obviously a little bit of code bloat with the class hierarchy.

This diff also removes the previous `local_result` and `receiver` kinds which were implemented with `NamedKind`. They were unused.

Reviewed By: anwesht

Differential Revision: D43072761

fbshipit-source-id: 4e3e801620179be80d4cd06120fad561af29314f