Use the taint representation to represent propagations [backward analysis 7/n]
Summary:
# Context
We currently use a representation for propagations (the information that the taint on an argument should be propagated to another argument or return value) that is different from the taint representation. The taint representation is basically a `TaintAccessPathTree` while propagations uses `PropagationAccessPathTree`.

# Problem
There are a few downsides with using a different representation:

1. The taint representation and the propagation representation needs to be consistent, which means updating one might require updating the other. This has some maintenance cost. Slight differences in the taint and propagation representation can also lead to bugs or imprecisions.
2. We are working on implementing a backward analysis, which will use the taint representation to propagate backward a `LocalReturn` and `LocalArgument(0)`. Once the analysis is complete, we could convert that representation into the representation for propagation, but this is just extra overhead.
3. We are also working on taint transforms, and we will need to represent that a propagation can "transform" a taint. This is not possible with the current representation for propagations, and would add extra overhead.
4. This is different from other static analyzers (e.g, pysa). Making our tools use the same representation has other benefits.

The only upside from using a different representation is performance. Our current representation for propagation is indeed more lightweight, since it does not use: the callee, callee port, location, origins, via-features, etc.

# Solution
Let's just use the same representation for taint and propagations, and monitor the impact on performance.

Reviewed By: anwesht

Differential Revision: D43203954

fbshipit-source-id: b0800a0cb504c841d1e69a9ad33c9e8b0481da3a