remove `add_features_to_arguments` for ContextFeature.json
Summary:
Context: We're trying to add propagation of `add_features_to_arguments`, and this diff came out of finding the biggest offenders of side-effect taint propagation. `via:context`, which is solely created by this generator, was by far the most commonly propagated feature. I added logging to the propagation of `add_features_to_arguments` that tracked the numbers.

The following was the number of propagations per feature:

```
[
  [
    21707218,
    "`via:context`"
  ],
  [
    19223222,
    "`via-stripped-logcat`"
  ],
  [
    11907627,
    "`via-canonical-path`"
  ],
  [
    9183242,
    "`via-post`"
  ],
  [
    9023322,
    "`cast:uri`"
  ],
  [
    8686352,
    "`via-secure-uri-parser`"
  ],
  [
    8455582,
    "`has-intent-extras`"
  ],
  [
    7332116,
    "`has-intent-data`"
  ],
  [
    7280761,
    "`via-securependingintent`"
  ],
  [
    7159963,
    "`via-explicit-intent`"
  ],
  [
    6159814,
    "`via-escape-html`"
  ],
  [
    6023925,
    "`has-intent-action`"
  ],
  [
    5788800,
    "`has-intent-category`"
  ],
  [
    2413277,
    "`via-strict-file`"
  ],
  [
    861057,
    "`via-permission-check`"
  ],
  [
    9721,
    "`via-secure-file-receiver`"
  ],
  [
    91,
    "`has-no-sensitive-data`"
  ],
  [
    2,
    "`via-certificate-pinner`"
  ]
]
```

In practice, we saw this feature get propagated all over the place. The side-effect part of the via:context is mostly a historical feature at this point. We might also consider replacing these with sanitizers at some point, but let's be conservative for now.

Reviewed By: GerbenJavado

Differential Revision: D43192995

fbshipit-source-id: d2f94eae6a42cd763b97c3fc93d567b7c7cc3b6c