Rename the forward taint analysis files [backward analysis 10/n]
Summary:
# Problem
We want to implement a backward analysis in Mariana Trench to improve precision.

The backward analysis will need to use the alias analysis (i.e, memory locations) for the same reason that the forward analysis does: to properly handle cases where multiple registers point to the same memory location.

Currently, memory locations are computed during the forward analysis, at the same time as the taint analysis.

# Solution
In this section, we will discuss different approaches to use memory locations in the backward analysis.

## Solution A
We could compute memory locations again during the backward analysis. This could probably work but seems harder to reason about: when executing instructions backward, that means we will need to create "placeholder" memory locations for arguments instead of creating memory locations for the result, like we do in the forward analysis. We could also have results different from the forward analysis.

## Solution B
We could compute memory locations in the forward analysis, and pass the result to the backward analysis.
The downside of this approach is that the forward analysis will be more complex and it might be better to separate concerns, hence solution C.

## Solution C
We could perform a forward alias analysis prior to the forward taint and backward taint analysis. This way, the forward and backward analysis will use the same logic to fetch memory locations, prenventing subtle bugs.

This will also allow us to cache the alias analysis result, so we do not recompute it each iteration.

# Decision
Here, we decided to move forward with the solution C, because of all the upsides and downsides mentioned above.

# This diff
In this diff, we rename the forward analysis into `ForwardTaint*` (`ForwardTaintEnvironment`, `ForwardTaintFixpoint`, `ForwardTaintTransfer`). In later diffs, we will introduce a `ForwardAlias*` that computes memory locations.

Reviewed By: anwesht

Differential Revision: D43260182

fbshipit-source-id: 22bc95490eaf6e6df6a630337badb3b94de59ae9