Add a class to store the alias analysis results [backward analysis 12/n]
Summary:
See context in D43260182 (https://github.com/facebook/mariana-trench/commit/808b4ae7647083a25f4c106c80824c14f5febee5).

In this diff, we introduce a class responsible for holding the results of the alias analysis (i.e, memory locations) so they can be used in both the forward and backward taint analysis.
This is built as a mapping from instructions to a class holding the environment, and the result memory locations of the instruction.
Note that it also holds the location of the instruction, so that we have consistent locations between the forward and backward analysis.

Note that the class will be used in a follow-up.

Reviewed By: sinancepel

Differential Revision: D43339368

fbshipit-source-id: 71df4452bf5a9c89fee8c161650ac7b52961ab6c