Move the alias analysis into its own fixpoint [backward analysis 13/n]
Summary:
See context in D43260182 (https://github.com/facebook/mariana-trench/commit/808b4ae7647083a25f4c106c80824c14f5febee5).

In this diff, we extract the alias analysis - i.e, building the memory location environment - from the forward taint analysis and move it into its own fixpoint. The fixpoint then writes its results into the `AliasAnalysisResults` stored in the method context.

Note that this diff is actually a no-op since the analysis is not used.
This is done in the next diff. This makes it easier to test for performance regressions.

Reviewed By: sinancepel

Differential Revision: D43339530

fbshipit-source-id: a7145c7d0787b9419156e8c4440673ccac663a4d