rename SignatureConstraint to be SignaturePatternConstraint and support parsing `signature_pattern`
Summary:
The longer term idea here is to support more concise constraints in Mariana Trench to hopefully stop relying so much on regex matching when we do our constraint matches.

For now, make a backwards-compatible change that'll eventually enable us to change the meaning of the `"signature"` constraint, and make it explicit that we support regexes.

Reviewed By: arthaud

Differential Revision: D43361283

fbshipit-source-id: ef51f048809b02911eb32db7f0a3779bd3ad0560