Use `PatriciaTreeSet::singleton()` instead of `size() == 1`
Summary: During our analysis, there are many cases where we check if a memory location set is a singleton. We currently call `size() == 1` followed by a `*set.elements().begin()`. This is suboptimal because calling `size()` is in `O(n)` in a patricia tree, and calling `begin()` might allocate memory. Let's use `singleton()` that we recently implemented in D43381839.

Reviewed By: anwesht

Differential Revision: D43413684

fbshipit-source-id: 7d12eeb3ccbb5f1c995fdfa0a470e4794c3d6c1d